Literature review and research
My project is still going in the trajectory mentioned in this document. I am currently experimenting with the limitations of NetLogo to see where I need to use Python to remedy those limitations. As I have mentioned in my project proposal, I will be extracting the data I collect via NetLogo and use Python to process it.