UML diagram and initial design
When I created this document, I didn't know a lot about coding in NetLogo. Therefore, I did the design in a pseudocode in a format that was more familiar to me, Python. However in my final verdict of which programming language to use for my project, I eventually decided to go with NetLogo. For one, this is a language I am not familiar with and I wanted to challenge myself and open myself to new opportunities and new learning feats. For two, I realized that there were ready to use templates for my simulation purposes that I can improve for my needs and use. These reasons combined got me to believe that I could accomplish my goals for this project using NetLogo.